backend/requirements.txt
Flask
openai
python-dotenv
Flask-Cors
supabase-py
requests
